requirements.txt
websockets
aiohttp
python-jose[cryptography]
motor
